Add heartbeating to detect chat network failures
Sometimes the underyling chat adapter gets stuck, or hipchat goes down
or disappears, and Lita still thinks things are good but it's not
actually online.

This change adds a new health check routine where the bot will periodically
message itself and measure the result to see if the chat system is healthy.
If it's not healthy, lita will restart itself.