Add 'publish' command.
This publish command largely behaves the same as the old jarvis (or is
intended to behave that way) with one exception - the specs are not run.

We will soon rely on Jenkins reports to determine if the tests have run
and are passing instead of running them on Jarvis.

Also added a Jarvis::execute method to help in running subprocesses and
logging the output.