Add a way to have commands log their current state.
This was a quick implementation using Cabin to provide a way for each
thread to log its current work and retrieve the last item for each
thread.

This exposes a new command `status` to jarvis which will emit each
thread's current state assuming one has logged it as such.

Fixes #48