backport: proof of concept
The idea is to apply the resulting diff of a PR as a single commit to
another branch.

The commit message would be an aggregation of all of the commit
messages, each with a reference to the original commit hash. The first
line of the commit message will indicate that it is a backport and will
reference the PR number.