Modify merge command to perform merge on target branch
The merge command has historically treated the list of branches as equals;
we would format-patch the change and then apply it to the list one by one.

This would mean that the github PR would not be marked as merged as we
were creating a new commit. This change modifies that, where the first
branch is merged with a normal git merge, and the remaining branches are
treated as before with 'git am'.

The new merge command will not append a "fixes #pr_number" sentence to
the commit message as it did before. Doing this would create a different
commit preventing the github PR from being marked as merged.

We keep the old command as "oldmerge".

This change also supports an empty branch list as an argument. In this
situation the PR's base branch will be used.

Resolves #54