Use github merge API for first target branch
To avoid "merge" commits when merging PRs that require rebase,
we rely on the github merge api instead of local "git merge" or
"git am" for the PR's target branch.

To do this we need to change the merge command flow a little bit.
This is to ensure that we first verify that the PR can be merged to the
target branch + all the backport branches, and only then make the api
call + git push.